TagLib Wrapper
