+++ NPlaylist/NPlaylist.Business/TagLib/ITagLibWrapper.cs
@@ -1,0 +1,9 @@
+﻿using TagLib;
+
+namespace NPlaylist.Business.TagLib
+{
+    public interface ITagLibWrapper
+    {
+        File Create(string filePath);
+    }
+}
