+++ NPlaylist/NPlaylist.Business/TagLib/TagLibWrapper.cs
@@ -1,0 +1,12 @@
+﻿using TagLib;
+
+namespace NPlaylist.Business.TagLib
+{
+    public class TagLibWrapper : ITagLibWrapper
+    {
+        public File Create(string filePath)
+        {
+            return File.Create(filePath);
+        }
+    }
+}
