+++ NPlaylist/NPlaylist.Infrastructure/System/DateTimeWrapper.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace NPlaylist.Infrastructure.System
+{
+    public class DateTimeWrapper : IDateTimeWrapper
+    {
+        public DateTime Now => DateTime.Now;
+    }
+}
