+++ NPlaylist/NPlaylist.Infrastructure/System/IPathWrapper.cs
@@ -1,0 +1,9 @@
+﻿namespace NPlaylist.Infrastructure.System
+{
+    public interface IPathWrapper
+    {
+        string GetDirectoryName(string filePath);
+
+        string GetExtension(string fileName);
+    }
+}
