--- NPlaylist/NPlaylist.Infrastructure/Wrappers/DateTimeWrapper/DateTimeWrapperImpl.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace NPlaylist.Infrastructure.Wrappers.DateTimeWrapper
-{
-    public class DateTimeWrapperImpl : IDateTimeWrapper
-    {
-        public DateTime Now => DateTime.Now;
-    }
-}
