--- NPlaylist/NPlaylist.Infrastructure/Wrappers/PathWrapper/IPathWrapper.cs
@@ -1,8 +1,0 @@
-﻿namespace NPlaylist.Infrastructure.Wrappers.PathWrapper
-{
-    public interface IPathWrapper
-    {
-        string GetExtension(string fileName);
-        string GetDirectoryName(string filePath);
-    }
-}
