+++ NPlaylist/NPlaylist.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 16.0.28803.156
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NPlaylist", "NPlaylist\NPlaylist.csproj", "{623E8331-03FC-46FD-AEF9-850A06C7E779}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NPlaylist.Infrastructure", "NPlaylist.Infrastructure\NPlaylist.Infrastructure.csproj", "{1273281D-EE85-4812-9A97-65C313DE3FC1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NPlaylist.Infrastructure", "NPlaylist.Infrastructure\NPlaylist.Infrastructure.csproj", "{1273281D-EE85-4812-9A97-65C313DE3FC1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NPlaylist.Business", "NPlaylist.Business\NPlaylist.Business.csproj", "{C5B4F259-1C52-40EB-B18D-0208F4985188}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NPlaylist.Persistance", "NPlaylist.Persistance\NPlaylist.Persistance.csproj", "{1B1791B9-8CD2-44F6-AC7E-9F0463F0CE88}"
 EndProject
@@ -15,7 +17,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	ProjectSection(SolutionItems) = preProject
 		db_migrate_and_update.txt = db_migrate_and_update.txt
 	EndProjectSection
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -38,6 +39,10 @@ Global
 		{B6E9A431-52B8-48CE-AC61-C1FFE793C75C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B6E9A431-52B8-48CE-AC61-C1FFE793C75C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B6E9A431-52B8-48CE-AC61-C1FFE793C75C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C5B4F259-1C52-40EB-B18D-0208F4985188}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C5B4F259-1C52-40EB-B18D-0208F4985188}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C5B4F259-1C52-40EB-B18D-0208F4985188}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C5B4F259-1C52-40EB-B18D-0208F4985188}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
