+++ NPlaylist/NPlaylist/Startup.cs
@@ -1,18 +1,15 @@
-﻿using System.IO;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Identity;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.FileProviders;
-using NPlaylist.Infrastructure.Wrappers.DateTimeWrapper;
-using NPlaylist.Infrastructure.Wrappers.DirectoryWrapper;
-using NPlaylist.Infrastructure.Wrappers.GuidWrapper;
-using NPlaylist.Infrastructure.Wrappers.PathWrapper;
 using NPlaylist.Authentication;
-using NPlaylist.Data;
+using NPlaylist.Business.TagLib;
+using NPlaylist.Infrastructure.System;
+using System.IO;
+using NPlaylist.Persistance;
 
 namespace NPlaylist
 {
@@ -24,36 +21,6 @@ namespace NPlaylist
         }
 
         public IConfiguration Configuration { get; }
-
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.Configure<CookiePolicyOptions>(options =>
-            {
-                options.CheckConsentNeeded = context => true;
-            });
-
-            services.AddSingleton<IFileProvider>(
-                new PhysicalFileProvider(
-                    Path.Combine(Directory.GetCurrentDirectory(), "wwwroot")));
-
-            ConfigureDbContexts(services);
-
-            services.Configure<IdentityOptions>(options =>
-            {
-                options.Password.RequireDigit = false;
-                options.Password.RequiredLength = 5;
-                options.Password.RequireLowercase = false;
-                options.Password.RequireNonAlphanumeric = false;
-                options.Password.RequireUppercase = false;
-            });
-
-            services.AddScoped<IDateTimeWrapper, DateTimeWrapperImpl>();
-            services.AddScoped<IDirectoryWrapper, DirectoryWrapperImpl>();
-            services.AddScoped<IGuidWrapper, GuidWrapperImpl>();
-            services.AddScoped<IPathWrapper, PathWrapperImpl>();
-
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
-        }
 
         public void Configure(IApplicationBuilder app, IHostingEnvironment env)
         {
@@ -81,6 +48,36 @@ namespace NPlaylist
             });
         }
 
+        public void ConfigureServices(IServiceCollection services)
+        {
+            services.Configure<CookiePolicyOptions>(options =>
+            {
+                options.CheckConsentNeeded = context => true;
+            });
+
+            services.AddSingleton<IFileProvider>(
+                new PhysicalFileProvider(
+                    Path.Combine(Directory.GetCurrentDirectory(), "wwwroot")));
+
+            ConfigureDbContexts(services);
+
+            services.Configure<IdentityOptions>(options =>
+            {
+                options.Password.RequireDigit = false;
+                options.Password.RequiredLength = 5;
+                options.Password.RequireLowercase = false;
+                options.Password.RequireNonAlphanumeric = false;
+                options.Password.RequireUppercase = false;
+            });
+
+            services.AddScoped<ITagLibWrapper, TagLibWrapper>();
+            services.AddScoped<IDateTimeWrapper, DateTimeWrapper>();
+            services.AddScoped<IDirectoryWrapper, DirectoryWrapper>();
+            services.AddScoped<IGuidWrapper, GuidWrapper>();
+            services.AddScoped<IPathWrapper, PathWrapper>();
+
+            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+        }
         private void ConfigureDbContexts(IServiceCollection services)
         {
             services.AddDbContext<ApplicationDbContext>();
